Refactor MongoConnector Code (#17)
* Added tar files to git commit and only adding trades with valid symbols to DB

* Refactored mongo connector

* Price is double instead of int now

* Missed a double conversion

+++ src/main/java/com/ibm/hybrid/cloud/sample/stocktrader/tradehistory/mongo/StockPurchase.java
@@ -11,7 +11,7 @@ public class StockPurchase implements Serializable{
     private String owner;
     private String symbol;
     private int shares;
-    private int price;
+    private double price;
     private String when;
     private int commission;
 
@@ -21,7 +21,7 @@ public class StockPurchase implements Serializable{
         this.owner = jObject.getString("owner");
         this.symbol = jObject.getString("symbol");
         this.shares = jObject.getInt("shares");
-        this.price = jObject.getInt("price");
+        this.price = jObject.getDouble("price");
         this.when = jObject.getString("when");
         this.commission = jObject.getInt("commission");
     }
@@ -42,7 +42,7 @@ public class StockPurchase implements Serializable{
         return shares;
     }
 
-    public int getPrice() {
+    public double getPrice() {
         return price;
     }
